packages/school/import/requirements.txt
pymongo[srv]==3.11.0
python-jwt==3.3.0
passlib==1.7.4
bcrypt==3.2.0
pandas==1.1.3
